requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.31
#mysqlclient==2.0.3
PyMySQL==1.0.2
Werkzeug==2.2.2
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
Flask-Migrate == 4.0.7
alembic==1.13.3
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
Flask==2.0.1
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
PyMySQL==1.0.2
python-dotenv==1.0.1
SQLAlchemy==1.4.31
typing_extensions==4.12.2
Werkzeug==2.2.2
WTForms==3.1.2
gunicorn==23.0.0
